requirements.txt
web3==6.13.0
pytest==7.4.3
pytest-mock==3.12.0
python-telegram-bot==13.8.1
flake8==6.1.0
SQLAlchemy==2.0.24
psycopg2-binary==2.9.9
retrying==1.3.4
alembic==1.13.1
